Use factory method of instantiating a class. Minor code fixes.
